Spostate regioni in una cartella
Vedere "data" o "css" in mezzo ai nomi di regioni mi disturbava
